feat: Reconnect to existing SAS Session if possible.
A sas session connection is made as needed by the extension.
This session is maintained throughout the vscode session. Many
things can cause the extension to reload (browser refresh, vscode reload, etc)
and this causes a new connection to be established, which creates a
brand new SAS session.

If an existing session is available, we should reconnect to it.

This is gated behind a REST profile option "reconnect"

Signed-off-by: Joseph Henry <Joseph.Henry@sas.com>