refactor: Session Log Streaming (#380)
Check the session job status before streaming the log to avoid various timing problems

Additionally, LibraryModel is currently creating a session. Modify LibraryModel to attach the log handler when LibraryModel creates a session in order to see the session log when LibraryModel invokes session setup. Fixes current issues where if the user opens the Library Panel first, before executing code, then the session log will not be output to the channel. Additionally, when running the actual sas file, it would also not be seen there, which is not desired behavior.