fix: check job results for success or error (#468)
Relaxes the check on compute job state to check for results even if the job errors out. This has the added benefit of showing results that are valid, in scenarios where part of the code produced results, but a subsequent part of code resulted in a compute job error state.