feat: add locale script (#464)
**Summary**
This adds a new locale script to aid in translating our plugin. The following commands are available
 - `npm run locale --new=<locale>`: Creates a new locale
 - `npm run locale --update-locale=<locale>`: updates a locale
 - `npm run locale --update-locales`: This updates all locales and is used internally in the `check-translations` github action

A new job, `check-translations`, was added to github actions as part of these changes. Basically, this fails github actions _if_ `npm run locale --update-locales` results in any changes. The idea here is that if we're adding a new feature, we don't want to merge it until we get it translated.

With that in mind, `check-translations` is not _required_, so if it fails, code can still be merged. In that case, it just serves as a visual reminder that something needs to happen in the future.

**Testing**
 - Ran `npm run locale --new=pt-br`
 - Translated some strings
 - Installed the language pack and made sure those strings were translated