feat: add libraries & tables for IOM/COM (#880)
**Summary**
This adds support for library/table viewing under IOM & COM connections. The general approach to getting the information we need is to:
 - Temporarily re-route session log output to a variable
 - Run some sql or a data step to get library & table info (similar to what EG does)
 - Parse data between a starting tag and end tag (since we don't care about _all_ input).

One somewhat hacky thing is parsing the SAS system line (On the en-us configuration I used, this said "The SAS System"). This line comes at the start of each new page of data, causing issues with data parsing. This is subject to change with any feedback in code review that may provide a better approach to solving this.

**Testing**
 - [x] Tested new functionality by starting with an empty set of profiles
 - [x] Tested viewing/refreshing libraries
 - [x] Tested viewing/refreshing tables
 - [x] Tested viewing table data (column types, row data, etc)
 - [x] Tested downloading data as a CSV
 - [x] Tested deleting tables
 - [x] Make sure log works as expected for regular sas program