fix: rehydrate result panels on restarts (#935)
Currently, Results from executed programs are not re-rendered if there are Result tabs open when vscode exits/restarts. On a restart, the editor group is left behind without any content. This leads to an undesirable user experience. The extension api supports persisting state for webviews that can be leveraged in a such a way that we can rehydrate the result html for these panels in between vscode exits/restarts.